Hawkes Processes Simulator

Warning

The documentation is still missing. Please see the "examples" section for usage.

HawkesSimulator.clear_trains!

HawkesSimulator.clear_trains! — Function

clear_trains!(ps::PopulationState,tlim=30.0)

Moves the spike train older than tlim to history. The spikes in history are not considered for interaction. This is necessary to make the dynamics efficient.

Examples

1D and 2D, exponential kernel

1D delayed-alpha kernel

2D delayed-alpha interactions, and non-delayed autapses