Hawkes Processes Simulator

Warning

The documentation is still missing. Please see the "examples" section for usage.

Examples

Simple case: 1D and 2D with exponential kernel

1D Hawkes process with a delayed kernel

Show that rate in Hawkes is equivalent to linear network, in a 2D model

Index

HawkesSimulator.draw_spike_raster

HawkesSimulator.dynamics_step_singlepopulation!

HawkesSimulator.flush_trains!

HawkesSimulator.warmup_step!

HawkesSimulator.draw_spike_raster — Method

 draw_spike_raster(trains::Vector{Vector{Float64}},
  dt::Real,Tend::Real;
  Tstart::Real=0.0,
  spike_size::Integer = 5,
  spike_separator::Integer = 1,
  background_color::Color=RGB(1.,1.,1.),
  spike_colors::Union{C,Vector{C}}=RGB(0.,0.0,0.0),
  max_size::Real=1E4) where C<:Color

Draws a matrix that contains the raster plot of the spike train.

Arguments

Trains : Vector of spike trains. The order of the vector corresponds to

the order of the plot. First element is at the top, second is second row, etc.

dt : time interval representing one horizontal pixel

Tend : final time to be considered

Optional arguments

Tstart : starting time

max_size : throws an error if image is larger than this number (in pixels)

spike_size : heigh of spike (in pixels)

spike_separator : space between spikes, and vertical padding

background_color : self-explanatory

spike_colors : if a single color, color of all spikes, if vector of colors,

color for each neuron (length should be same as number of neurons)

returns

raster_matrix::Matrix{Color} you can save it as a png file

HawkesSimulator.dynamics_step_singlepopulation! — Method

dynamics_step_singlepopulation!(t_now::Real,ntw::RecurrentNetwork)

Iterates a one-population network up until its next spike time. This is done by computing a next spike proposal for each neuron, and then picking the one that happens sooner. This spike is then added to the spiketrain for that neuron. The fundtion returns the new current time of the simulation.

For long simulations, this functions should be called jointly with flush_trains!. Otherwise the spike trains will keep growing, making the propagation of signals extremely cumbersome.

Arguments

t_now - Current time of the simulation

ntw - The network

Returns

t_now_new - the new current time of the simulation

HawkesSimulator.flush_trains! — Method

flush_trains!(ps::PopulationState,Ttrigger::Real;
    Tflush::Union{Real,Nothing}=nothing)

Spike history is spiketimes that do not interact with the kernel (because too old) This function compares most recent spike with spike history, if enough time has passed (measured by Ttrigger) it flushes the spiketrain up to Tflush into the history.

HawkesSimulator.warmup_step! — Method

warmupstep!(tnow::Real,ntw::RecurrentNetwork, warmuprates::Union{Vector{Float64},Vector{Vector{Float64}}}) -> tend

In the warmup phase, all neurons fire as independent Poisson process with a rate set by warmup_rates. This is useful to quick-start the network, or set itinitial conditions that are far from the stable point.

Arguments

t_now - current time

ntw - the network (warning: weights and kernels are entirely ignored here)

warmup_rates - the desired stationary rates. In a one-population network, it is a vector with the desired rates. In a multi-population network, is a collection of vectors, where each vector refers to one population.

TODO : add other file. [comment]: <> ( 1. 2D delayed-alpha interactions, and non-delayed autapses 1. 1D delayed-alpha kernel 1. Plasticity examples, with pairwise STPD and triplets )