Hawkes Processes Simulator

Warning

The documentation is still missing. Please see the "examples" section for usage.

Examples

Simple case: 1D and 2D with exponential kernel

1D Hawkes process with a delayed kernel

Show that rate in Hawkes is equivalent to linear network, in a 2D model

Plasticity, pairwise STDP and triplet-based

Plasticity, rate-dependent STDP

Index

HawkesSimulator.bin_spikes

HawkesSimulator.draw_spike_raster

HawkesSimulator.dynamics_step_singlepopulation!

HawkesSimulator.flush_trains!

HawkesSimulator.get_trains

HawkesSimulator.numerical_rates

HawkesSimulator.warmup_step!

HawkesSimulator.bin_spikes — Method

bin_spikes(Y::Vector{R},dt::R,Tend::R;Tstart::R=0.0) where R

Arguments

Y::Vector{<:Real} : vector of spike times

dt::Real : time bin size

Tend::Real : end time for the raster

Optional argument

Tstart::Real=0.0 : start time for the raster

Returns

binned_spikes::Vector{<:Integer} : binned_spikes[k] is the number of spikes that occur in the timebin k (i.e. between Tstart + (k-1)*dt and Tstart + k*dt)

HawkesSimulator.draw_spike_raster — Method

drawspikeraster(trains::Vector{Vector{Float64}}, dt::Real,Tend::Real; Tstart::Real=0.0, spikesize::Integer = 5, spikeseparator::Integer = 1, backgroundcolor::Color=RGB(1.,1.,1.), spikecolors::Union{C,Vector{C}}=RGB(0.,0.0,0.0), max_size::Real=1E4) where C<:Color

Draws a matrix that contains the raster plot of the spike train.

Arguments

Trains : Vector of spike trains. The order of the vector corresponds to the order of the plot. First element is at the top, second is second row, etc.

dt : time interval representing one horizontal pixel

Tend : final time to be considered

Optional arguments

Tstart::Real : starting time

spike_size::Integer : heigh of spike (in pixels)

spike_separator::Integer : space between spikes, and vertical padding

background_color::Color : self-explanatory

spike_colors::Union{Color,Vector{Color}} : if a single color, color of all spikes, if vector of colors, color for each neuron (length should be same as number of neurons)

max_size::Integer : throws an error if image is larger than this number (in pixels)

Returns

raster_matrix::Matrix{Color} you can save it as a png file

HawkesSimulator.dynamics_step_singlepopulation! — Method

dynamics_step_singlepopulation!(t_now::Real,ntw::RecurrentNetwork)

Iterates a one-population network up until its next spike time. This is done by computing a next spike proposal for each neuron, and then picking the one that happens sooner. This spike is then added to the spiketrain for that neuron. The fundtion returns the new current time of the simulation.

For long simulations, this functions should be called jointly with flush_trains!. Otherwise the spike trains will keep growing, making the propagation of signals extremely cumbersome.

Arguments

t_now - Current time of the simulation

ntw - The network

Returns

t_now_new - the new current time of the simulation

HawkesSimulator.flush_trains! — Method

flush_trains!(ps::PopulationState,Ttrigger::Real;
    Tflush::Union{Real,Nothing}=nothing)

Spike history is spiketimes that do not interact with the kernel (because too old) This function compares most recent spike with spike history, if enough time has passed (measured by Ttrigger) it flushes the spiketrain up to Tflush into the history.

HawkesSimulator.get_trains — Method

function gettrains(rec::RecFullTrain{N}; Nneus::Integer = 0, popidx::Integer = 1) where N

Read full recorded train from the population selected by pop_idx (default 1).

HawkesSimulator.numerical_rates — Method

numericalrates(spikesrecorder,Nneus,Tend;pop_idx=1,Tstart=0.0)

Returns the numerical rates of the neurons in the population pop_idx that has been recorded through spikes_recorder (a RecFullTrain or a RecFullTrainContent object). The rates are computed in the time interval [Tstart,Tend] and are returned as a vector of length Nneus.

HawkesSimulator.warmup_step! — Method

warmupstep!(tnow::Real,ntw::RecurrentNetwork, warmuprates::Union{Vector{Float64},Vector{Vector{Float64}}}) -> tend

In the warmup phase, all neurons fire as independent Poisson process with a rate set by warmup_rates. This is useful to quick-start the network, or set itinitial conditions that are far from the stable point.

Arguments

t_now - current time

ntw - the network (warning: weights and kernels are entirely ignored here)

warmup_rates - the desired stationary rates. In a one-population network, it is a vector with the desired rates. In a multi-population network, is a collection of vectors, where each vector refers to one population.